+++ skltn/CssBuilder.php
@@ -1,9 +1,19 @@
 <?php
-/*  AUTO GENERATED FILE - DO NOT EDIT !!
-    WP-SKELEKTON | MIT X11 License | https://github.com/AndiDittrich/WP-Skeleton
-    ------------------------------------
-    Generate Dynamic CSS Files
-*/
+// ---------------------------------------------------------------------------------------------------------------
+// -- WP-SKELETON AUTO GENERATED FILE - DO NOT EDIT !!!
+// --
+// -- Copyright (c) 2016-2018 Andi Dittrich
+// -- https://github.com/AndiDittrich/WP-Skeleton
+// --
+// ---------------------------------------------------------------------------------------------------------------
+// --
+// -- This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0.
+// -- If a copy of the MPL was not distributed with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
+// --
+// ---------------------------------------------------------------------------------------------------------------
+
+// Generate Dynamic CSS Files
+
 namespace Tweakr\skltn;
 
 class CssBuilder{
@@ -19,7 +29,13 @@ class CssBuilder{
 
     // add new ruleset
     public function add($selector, $rules = array()){
-        $this->_buffer[$selector] = $rules;
+        // selector already used ?
+        if (isset($this->_buffer[$selector])){
+            // merge
+            $this->_buffer[$selector] = array_merge($this->_buffer[$selector], $rules);
+        }else{
+            $this->_buffer[$selector] = $rules;
+        }
     }
 
     // add raw css
@@ -30,7 +46,7 @@ class CssBuilder{
     // render ruleset
     public function render(){
         // local output buffer
-        $css = '/* Tweakr Dynamical Generated Stylesheet - DO NOT EDIT */';
+        $css = '/* Tweakr dynamic generated stylesheet - DO NOT EDIT */';
 
         // get rulesets
         foreach ($this->_buffer as $selector => $rules){