+++ skltn/RewriteRuleHelper.php
@@ -1,10 +1,19 @@
 <?php
+// ---------------------------------------------------------------------------------------------------------------
+// -- WP-SKELETON AUTO GENERATED FILE - DO NOT EDIT !!!
+// --
+// -- Copyright (c) 2016-2018 Andi Dittrich
+// -- https://github.com/AndiDittrich/WP-Skeleton
+// --
+// ---------------------------------------------------------------------------------------------------------------
+// --
+// -- This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0.
+// -- If a copy of the MPL was not distributed with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
+// --
+// ---------------------------------------------------------------------------------------------------------------
 
-/*  AUTO GENERATED FILE - DO NOT EDIT !!
-    WP-SKELEKTON | MIT X11 License | https://github.com/AndiDittrich/WP-Skeleton
-    ------------------------------------
-    Filters the rewrite rules to allow an easier manipulation of them especially on plugin activation/deactivation
-*/
+// Filters the rewrite rules to allow an easier manipulation of them especially on plugin activation/deactivation
+
 namespace Tweakr\skltn;
 
 class RewriteRuleHelper{